+++ crates/scheduler/test.sh
@@ -15,9 +15,10 @@
 
 set -ex
 
+touch ../../target/applet.wasm
 cargo check --features=std
 cargo check --features=std,log
 cargo check --target=thumbv7em-none-eabi
 cargo check --target=thumbv7em-none-eabi --features=defmt
 cargo fmt -- --check
-cargo clippy -- --deny=warnings
+cargo clippy --features=std -- --deny=warnings